requirements.txt
numpy
pandas
scikit-learn
python-Levenshtein
umap-learn==0.5.3